requirements.txt
bidict==0.22.1
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==3.0.0
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
simple-websocket==1.0.0
Werkzeug==3.0.1
wsproto==1.2.0
